+++ test/RuleIdentityRegistry/Ownable/RuleIdentityRegistryOwnable2Step.t.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.20;
 import {Test} from "forge-std/Test.sol";
 import {HelperContract} from "../../HelperContract.sol";
 import {Ownable2StepTestBase, IOwnable2StepLike} from "../../utils/Ownable2StepTestBase.sol";
-import {IdentityRegistryMock} from "../../utils/IdentityRegistryMock.sol";
+import {IdentityRegistryMock} from "src/mocks/IdentityRegistryMock.sol";
 import {RuleIdentityRegistryOwnable2Step} from
     "src/rules/validation/deployment/RuleIdentityRegistryOwnable2Step.sol";